feat: modularize model and optimization extension APIs
