+++ voicehub/diffusion_serving/capabilities.py
@@ -13,7 +13,7 @@ from enum import Enum
 from typing import TYPE_CHECKING
 
 from voicehub.errors import VoiceHubError
-from voicehub.registry import normalize_model_type
+from voicehub.models.registry import normalize_model_type
 
 if TYPE_CHECKING:
     from voicehub.diffusion_serving.vllm_omni import VLLMOmniDiffusionPlugin
